front: change & refacto map popup
