+++ front/src/styles/scss/common/_common.scss
@@ -1,0 +1,1 @@
+@import 'map/popup'
