+++ front/src/styles/styles.scss
@@ -3,4 +3,5 @@
 @import 'scss/body';
 @import 'scss/home';
 @import 'scss/applications/operationalStudies';
-@import 'scss/common/components'
+@import 'scss/common/common';
+@import 'scss/common/components';
